Corrections

+++ src/MCA7/AutoSell/Main.php
@@ -19,8 +19,8 @@ use MCA7\AutoSell\provider\EconomyAPIProvider;
 use MCA7\AutoSell\provider\EconomyProvider;
 
 /*	
-	Credits to cosmicnebula200 for multiple economy provider integration 
-	https://github.com/cosmicnebula200/SellMe
+    Credits to cosmicnebula200 for multiple economy provider integration 
+    https://github.com/cosmicnebula200/SellMe
 */
 
 class Main extends PluginBase implements Listener
@@ -40,41 +40,41 @@ class Main extends PluginBase implements Listener
 		$this->getServer()->getPluginManager()->registerEvents($this, $this);
 		$this->blocks = $this->prices->getAll();
 		$this->economyProvider = match (strtolower($this->getConfig()->get('economy-provider')))
-        {
-            "bedrockeconomy" => new BedrockEconomyProvider(),
-            "capital" => new CapitalEconomyProvider(),
-            "economyapi" => new EconomyAPIProvider(),
-            default => null
-        };
+        	{
+            		"bedrockeconomy" => new BedrockEconomyProvider(),
+            		"capital" => new CapitalEconomyProvider(),
+            		"economyapi" => new EconomyAPIProvider(),
+            		default => null
+        	};
 
-        if ($this->economyProvider == null)
-        {
-            $this->yeet($this->getConfig()->get('economy-provider'));
-            return;
-        }
-        if (!$this->economyProvider->checkClass())
-        {
-            $this->yeet($this->economyProvider->getName());
-            return;
-        }
+        	if ($this->economyProvider == null)
+        	{
+            		$this->yeet($this->getConfig()->get('economy-provider'));
+            		return;
+        	}
+        	if (!$this->economyProvider->checkClass())
+        	{
+            		$this->yeet($this->economyProvider->getName());
+            		return;
+        	}
 	}
 
 
 	/**
-     * @return EconomyProvider|null
-     */
+     	* @return EconomyProvider|null
+     	*/
 
-    public function getEconomyProvider(): ?EconomyProvider
-    {
-        return $this->economyProvider;
-    }
+    	public function getEconomyProvider(): ?EconomyProvider
+    	{
+        	return $this->economyProvider;
+    	}
 
 
 	private function yeet(string $name): void
-    {
+    	{
         $this->getServer()->getLogger()->error("The respected class for the Economy Provider $name has not been found");
         $this->getServer()->getPluginManager()->disablePlugin($this);
-    }
+    	}
 
 
 	public function onLoad(): void
@@ -99,9 +99,7 @@ class Main extends PluginBase implements Listener
 
 	public function onJoin(PlayerJoinEvent $event): void
 	{
-
 		$player = $event->getPlayer()->getName();
-
 		if (!$this->db->getNested($player)) {
 			$this->db->setNested($player, "off");
 			$this->db->save();
@@ -241,7 +239,7 @@ class Main extends PluginBase implements Listener
 		if (isset($this->blocks[$drop->getName()])) {
 			$event->setDrops([]);
 			$itemname = $drop->getName();
-			$price = (float)$this->blocks[$itemname] * $count;
+			$price = (int)$this->blocks[$itemname] * $count;
 			$player->sendTip(
 				TextFormat::GREEN . "Sold" . TextFormat::AQUA . " " . $itemname . "x" . $count . TextFormat::GREEN . " for" . TextFormat::YELLOW . " $" . $price
 			);
